In a loop, set the direction of the 'start belt' correctly
Previously, the starting belt (the last one we found) was simply reversed.
In a loop, this was often not the expected behaviour. Instead, the start
belt should be set the same as everything else, opposite the direction of
the belt that feeds it. Now we remember this value when we detect the loop.