Fix bug with loops where initial selection was underground belt
If the start of the belt was determined to be the output of a connected underground belt,
an error would be raised due to the way the reversing code worked. To fix, we always set
the start of the belt to be the input of the underground belt if it is connected.