Flip belt lines when reversing belts
When a belt is reversed, the lines of items on the belt are also switched. This meanse that
if only one side of the belt is full, that side will 'jump' to the other side. These changes
make sure they stay on their original side.

This also ensures items on the underground belts are saved and switched correctly.

Thanks to Articulating for the core code and helping debug issues.