Fix crash when loader has no 'last_user'
When we reverse a loader, we have to remove and recreate it.
In some situations it's possible for an entity to not have a last_user,
probably when the upgrade planner is used.
If the old loader didn't have a last_user, we would try and set the last_user
on the new loader to that none value, which would crash.
We now check to make sure it exists before trying to set it.